requirements.txt
asgiref==3.8.1
bleach==6.2.0
boto3==1.34.34
botocore==1.34.162
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==2023.10
cryptography==44.0.2
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.0.2
django-admin-interface==0.25.0
django-admin-rangefilter==0.12.0
django-admin-thumbnails==0.2.9
django-allauth==0.60.1
django-ckeditor==6.7.1
django-colorfield==0.13.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-import-export==3.3.4
django-js-asset==3.1.2
django-jsonfield==1.4.1
django-simple-history==3.4.0
django-storages==1.14.2
django-summernote==0.8.20.0
django-taggit==5.0.1
django-tinymce==3.6.0
et_xmlfile==2.0.0
idna==3.10
jmespath==1.0.1
MarkupPy==1.18
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.4
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tablib==3.5.0
text-unidecode==1.3
urllib3==2.4.0
webencodings==0.5.1
wheel==0.45.1
xlrd==2.0.1
xlwt==1.3.0
